refactor: adjust value to be grater than 0 in transaction input

+++ models/transactions.go
@@ -15,8 +15,8 @@ type Transactions struct {
 }
 
 type TransactionsInput struct {
-	Description string `json:"description" binding:"required,max=50"`
-	Value       int64  `json:"value" binding:"required,gt=0"`
+	Description string  `json:"description" binding:"required,max=50"`
+	Value       float64 `json:"value" binding:"required,gt=0"`
 }
 
 func (t *Transactions) BeforeCreate(d *gorm.DB) (err error) {
